+++ lib/config/presets/internal/monorepos.ts
